+++ Jx.Cms.Themes/Vm/CatalogueVm.cs
@@ -33,5 +33,5 @@ public class CatalogueVm
     /// <summary>
     /// 页码信息
     /// </summary>
-    public Dictionary<string, int> Pagination { get; set; }
+    public Dictionary<string, long> Pagination { get; set; }
 }
